% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_intersect_left.R
\name{bed_intersect_left}
\alias{bed_intersect_left}
\title{Left join intersection of two tibble BEDs}
\usage{
bed_intersect_left(
  bedA,
  bedB,
  suffix = c("", ""),
  keepBcoords = T,
  keepBmetadata = T,
  suffixMetadataCols = F
)
}
\arguments{
\item{bedA}{A tibble BED with columns: chrom, start, end, ...(metadata cols)...}

\item{bedB}{A tibble BED with columns: chrom, start, end, ...(metadata cols)...}

\item{suffix}{A character vector of suffixes to add to metadata column names. (If the two input tibble BEDs share any common metadata column names, suffixes must be provided.)}

\item{keepBcoords}{If TRUE, coordinates from bedB are retained in the metadata. Must provide unique suffix for bedB to avoid start/end column duplication. Default is TRUE.}

\item{keepBmetadata}{If FALSE, bedB columns are dropped and only the original bedA columns are returned, filtered to those intervals that intersect with any bedB intervals.}

\item{suffixMetadataCols}{If TRUE, suffixes are also added to the end of metadata column names, not just the coordinate columns. Default is FALSE.}
}
\value{
A tibble BED of all intervals from bedA that contain intersects with bedB, plus all metadata columns from both
}
\description{
Intersect two tibble BEDs and retain the full intervals from the lefthand tibble BED that have any overlap with the righthand tibble BED.
Optionally retain the interval coordinates for the righthand tibble in the metadata.
If the two input tibble BEDs share any common metadata column names, or if keepBcoords = TRUE, then unique suffixes must be provided.
This function is built as a modification of the valr::bed_intersect() function.
Equivalent to bedtools -a A.bed -b B.bed -wa
}
