% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_BED.R
\name{import_BED}
\alias{import_BED}
\title{Import a BED}
\usage{
import_BED(bedfile, metadata_cols = NULL)
}
\arguments{
\item{bedfile}{A tab-delimited file in bed format with columns: chrom, start, stop, ...(metadata cols)...}

\item{metadata_cols}{A character vector of the names of metadata columns (bedfile cols 4+) in the BED file, to be imported as metadata in the tibble BED. These must be in the same order as they appear in the file, starting from column 4.}
}
\value{
A tibble representation of the BED file, optionally with metadata columns. If no metadata column names are provided, only the chrom, start and stop columns will be returned.
}
\description{
Import a BED file as a tibble
}
