% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_BEDPE_to_List.R
\name{import_BEDPE_to_List}
\alias{import_BEDPE_to_List}
\title{Import a BEDPE}
\usage{
import_BEDPE_to_List(
  bedpefile,
  metadata_cols = NULL,
  prefix_InteractionID = NULL
)
}
\arguments{
\item{bedpefile}{A tab-delimited file in bedpe format with columns: chromA, startA, endA, chromB, startB, endB, ...(metadata cols)...}

\item{metadata_cols}{A character vector of the names of metadata columns (bedfile cols 4+) in the BED file}
}
\value{
A paired (first + last) list representation of the BEDPE file as two valr-compatible tibble BEDs with an ID column matching pairs, optionally with metadata columns
}
\description{
Import a paired-end BEDPE file as a list object c(tibble BED of first end, tibble BED of last end).
The direction / grouping of first and last ends is arbitrary.
}
