% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_BEDPE.R
\name{intersect_BEDPE}
\alias{intersect_BEDPE}
\title{Intersect BEDPE list}
\usage{
intersect_BEDPE(bedpe, SNPend, TSSend)
}
\arguments{
\item{bedpe}{A tibble BEDPE list object representing chromosomal interaction looping data}

\item{SNPend}{A tibble BED of trait SNPs}

\item{TSSend}{A tibble BED of gene TSSs}
}
\value{
A tibble BED of all interaction loops in which one end intersects with a SNP and the other end intersects with a TSS. The coordinates for the SNP
}
\description{
Intersect BEDPE list object with SNPs and TSSs
}
